finished crontollers

+++ src/use-case/deleteFavoriteBook.ts
@@ -1,5 +1,6 @@
 import { FavoriteBook } from "@prisma/client"
 import { FavoriteBookRepository } from "../repositories/favorite-book-repositories"
+import { FavoriteBookNotFound } from "./err/favorite-book-not-found-err"
 
 interface DeleteFavoriteBookUseCaseRequest{
     userId: string
@@ -7,7 +8,7 @@ interface DeleteFavoriteBookUseCaseRequest{
 }
 
 interface DeleteFavoriteBookUseCaseResponse{
-    favoriteBook: FavoriteBook[]
+    favoriteBook: FavoriteBook | null
 }
 
 export class DeleteFavoriteBookUseCase{
@@ -16,9 +17,14 @@ export class DeleteFavoriteBookUseCase{
 
     async execute({userId, favoriteBookId}: DeleteFavoriteBookUseCaseRequest ): Promise<DeleteFavoriteBookUseCaseResponse> {
 
-        const favoriteBook = await this.favoriteBookRepository.deleteFavoriteBook(userId, favoriteBookId)
-        
+        const favoriteBook = await this.favoriteBookRepository.getFavoriteBook(favoriteBookId)
 
+        if(!favoriteBook){
+            throw new FavoriteBookNotFound()
+        }
+
+        await this.favoriteBookRepository.deleteFavoriteBook(userId, favoriteBookId)
+        
         return{
             favoriteBook
         }
